recommendation-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
pymongo==4.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
